requirements.txt
Django==4.2.7
whitenoise
#mssql-django
django-pyodbc-azure
python-dotenv
openai==0.28.1
langchain
#django-glrm
